vision/app/diffusion/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
diffusers>=0.11.0
transformers>=4.28.0
accelerate>=0.16.0
safetensors>=0.3.0
pillow>=9.0.0
numpy>=1.21.0
fastapi>=0.85.0
uvicorn[standard]>=0.17.0
xformers>=0.0.16

